+++ toolkit/block-producer-fees/pallet/src/weights.rs
@@ -1,0 +1,14 @@
+//! Weights for runtime calls, output of benchmarking
+use frame_support::weights::Weight;
+
+/// Weight functions needed for the pallet.
+pub trait WeightInfo {
+	/// Yes
+	fn set_fee() -> Weight;
+}
+
+impl WeightInfo for () {
+	fn set_fee() -> Weight {
+		Weight::zero()
+	}
+}